requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
jose==1.0.0
passlib==1.7.4
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==3.10.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
